+++ scanpipe/management/commands/__init__.py
@@ -144,7 +144,7 @@ class AddInputCommandMixin:
 
     def handle_input_files(self, inputs_files):
         """
-        Copies provided `inputs_files` to the project's `input` directory.
+        Copy provided `inputs_files` to the project's `input` directory.
         """
         copied = []
 
@@ -162,7 +162,7 @@ class AddInputCommandMixin:
     @staticmethod
     def validate_input_files(inputs_files):
         """
-        Raises an error if one of the provided `inputs_files` is not an existing
+        Raise an error if one of the provided `inputs_files` is not an existing
         file.
         """
         for file_location in inputs_files:
@@ -172,7 +172,7 @@ class AddInputCommandMixin:
 
     def handle_input_urls(self, input_urls):
         """
-        Fetches provided `input_urls` and stores it in the project's `input`
+        Fetch provided `input_urls` and stores it in the project's `input`
         directory.
         """
         downloads, errors = fetch_urls(input_urls)
@@ -192,7 +192,7 @@ class AddInputCommandMixin:
 
 def validate_input_files(file_locations):
     """
-    Raises an error if one of the provided `file_locations` is not an existing file.
+    Raise an error if one of the provided `file_locations` is not an existing file.
     """
     for file_location in file_locations:
         file_path = Path(file_location)
@@ -202,7 +202,7 @@ def validate_input_files(file_locations):
 
 def validate_pipelines(pipeline_names):
     """
-    Raises an error if one of the `pipeline_names` is not available.
+    Raise an error if one of the `pipeline_names` is not available.
     """
     for pipeline_name in pipeline_names:
         if pipeline_name not in scanpipe_app.pipelines: