+++ scanpipe/pipelines/inspect_manifest.py
@@ -31,7 +31,8 @@ class InspectManifest(Pipeline):
 
     Supports:
     - PyPI "requirements.txt" files
-    - SPDX document as JSON ".spdx.json"
+    - SPDX document as JSON ".spdx.json" files
+    - CycloneDX BOM as JSON ".bom.json" and ".cdx.json" files
     - AboutCode ".ABOUT" files
     """
 
@@ -44,7 +45,7 @@ class InspectManifest(Pipeline):
 
     def get_manifest_inputs(self):
         """
-        Locates all the manifest files from the project's input/ directory.
+        Locate all the manifest files from the project's input/ directory.
         """
         self.input_locations = [
             str(input.absolute()) for input in self.project.inputs()
@@ -52,7 +53,7 @@ class InspectManifest(Pipeline):
 
     def create_packages_from_manifest(self):
         """
-        Resolves manifest files into packages.
+        Resolve manifest files into packages.
         """
         for input_location in self.input_locations:
             default_package_type = resolve.get_default_package_type(input_location)