+++ scanpipe/views.py
@@ -65,7 +65,6 @@ from scanpipe.models import Project
 from scanpipe.models import ProjectError
 from scanpipe.models import Run
 from scanpipe.models import RunInProgressError
-from scanpipe.pipes import codebase
 from scanpipe.pipes import count_group_by
 from scanpipe.pipes import output
 
@@ -193,7 +192,7 @@ class TabSetMixin:
 
     def get_tabset_data(self):
         """
-        Returns the tabset data structure used in template rendering.
+        Return the tabset data structure used in template rendering.
         """
         tabset_data = {}
 
@@ -210,7 +209,7 @@ class TabSetMixin:
 
     def get_fields_data(self, fields):
         """
-        Returns the tab fields including their values for display.
+        Return the tab fields including their values for display.
         """
         fields_data = {}
 
@@ -235,7 +234,7 @@ class TabSetMixin:
 
     def get_field_value(self, field_name, render_func=None):
         """
-        Returns the formatted value for the given `field_name` on the current object.
+        Return the formatted value for the given `field_name` on the current object.
         """
         field_value = getattr(self.object, field_name, None)
 
@@ -250,7 +249,7 @@ class TabSetMixin:
     @staticmethod
     def get_field_label(field_name):
         """
-        Returns a formatted label for display based on the `field_name`.
+        Return a formatted label for display based on the `field_name`.
         """
         return field_name.replace("_", " ").capitalize().replace("url", "URL")
 
@@ -279,7 +278,7 @@ class TableColumnsMixin:
 
     def get_columns_data(self):
         """
-        Returns the columns data structure used in template rendering.
+        Return the columns data structure used in template rendering.
         """
         columns_data = []
 
@@ -326,7 +325,7 @@ class TableColumnsMixin:
     @staticmethod
     def get_field_label(field_name):
         """
-        Returns a formatted label for display based on the `field_name`.
+        Return a formatted label for display based on the `field_name`.
         """
         return field_name.replace("_", " ").capitalize().replace("url", "URL")
 
@@ -338,7 +337,7 @@ class TableColumnsMixin:
 
 class ExportXLSXMixin:
     """
-    Adds the ability to export the current filtered QuerySet of a `FilterView` into
+    Add the ability to export the current filtered QuerySet of a `FilterView` into
     the XLSX format.
     """
 
@@ -374,7 +373,7 @@ class ExportXLSXMixin:
 
 class PaginatedFilterView(FilterView):
     """
-    Adds a `url_params_without_page` value in the template context to include the
+    Add a `url_params_without_page` value in the template context to include the
     current filtering in the pagination.
     """
 
@@ -733,19 +732,6 @@ class ProjectResetView(ConditionalLoginRequired, ProjectViewMixin, generic.Delet
         return redirect(project)
 
 
-class ProjectTreeView(ConditionalLoginRequired, ProjectViewMixin, generic.DetailView):
-    template_name = "scanpipe/project_tree.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        fields = ["name", "path"]
-        project_codebase = codebase.ProjectCodebase(self.object)
-        context["tree_data"] = [codebase.get_tree(project_codebase.root, fields)]
-
-        return context
-
-
 @conditional_login_required
 def execute_pipeline_view(request, uuid, run_uuid):
     project = get_object_or_404(Project, uuid=uuid)
@@ -787,7 +773,7 @@ def delete_pipeline_view(request, uuid, run_uuid):
 
 def project_results_json_response(project, as_attachment=False):
     """
-    Returns the results as JSON compatible with ScanCode data format.
+    Return the results as JSON compatible with ScanCode data format.
     The content is returned as a stream of JSON content using the JSONResultsGenerator
     class.
     If `as_attachment` is True, the response will force the download of the file.
@@ -959,6 +945,8 @@ class CodebaseResourceDetailsView(
     ConditionalLoginRequired, ProjectRelatedViewMixin, generic.DetailView
 ):
     model = CodebaseResource
+    slug_field = "path"
+    slug_url_kwarg = "path"
     template_name = "scanpipe/resource_detail.html"
     annotation_types = {
         CodebaseResource.Compliance.OK: "ok",
@@ -972,8 +960,8 @@ class CodebaseResourceDetailsView(
     @staticmethod
     def get_annotation_text(entry, field_name, value_key):
         """
-        A workaround to get the license_expression until the data structure is
-        updated on the ScanCode-toolkit side.
+        Get the license_expression until the data structure is updated on the
+        ScanCode-toolkit side.
         https://github.com/nexB/scancode-results-analyzer/blob/6c132bc20153d5c96929c
         f378bd0f06d83db9005/src/results_analyze/analyzer_plugin.py#L131-L198
         """
@@ -1191,6 +1179,8 @@ class CodebaseResourceRawView(
     generic.base.View,
 ):
     model = CodebaseResource
+    slug_field = "path"
+    slug_url_kwarg = "path"
 
     def get(self, request, *args, **kwargs):
         resource = self.get_object()