+++ CHANGELOG.rst
@@ -4,11 +4,29 @@ Changelog
 v33.0.0 (unreleased)
 --------------------
 
+- Add support for CycloneDX 1.4 to the "inspect-manifest" pipeline to import SBOM into
+  a Project.
+  https://github.com/nexB/scancode.io/issues/583
+
 - Refine the "Command Line Interface" documentation about the ``scanpipe`` command
   usages in the Docker context.
-  Add the /app workdir in the PYTHONPATH env of the Docker file to make the ``scanpipe``
-  entry point available while running ``docker compose`` commands.
+  Add the /app workdir in the "PYTHONPATH" env of the Docker file to make the
+  ``scanpipe`` entry point available while running ``docker compose`` commands.
   https://github.com/nexB/scancode.io/issues/616
+
+- Add new tutorial about the "find vulnerabilities" pipeline and the vulnerablecode
+  integration in the documentation.
+  https://github.com/nexB/scancode.io/issues/600
+
+- Use CodebaseResource ``path`` instead of ``id`` as slug_field in URL navigation.
+  https://github.com/nexB/scancode.io/issues/242
+
+- Remove dead code related to the project_tree view
+  https://github.com/nexB/scancode.io/issues/623
+
+- Update ``scanpipe.pipes.ProjectCodebase`` and related code to work properly
+  with current Project/CodebaseResource path scheme.
+  https://github.com/nexB/scancode.io/pull/624
 
 v32.0.1 (2023-02-20)
 --------------------